Update documentation and changelog
