+++ rustler/src/resource/error.rs
@@ -1,0 +1,3 @@
+/// Indicates that a resource has not been registered successfully
+#[derive(Clone, Copy, Debug)]
+pub struct ResourceInitError;
